pyannote-api/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pyannote.audio==3.3.0
torch==2.2.0
python-multipart==0.0.6
